+++ src/main/java/net/deathpole/deathbot/Services/Impl/ChartServiceImpl.java
@@ -529,7 +529,7 @@ public class ChartServiceImpl implements IChartService {
 
     @Override
     public HashMap<LocalDateTime, BigDecimal> getSRStatsForPlayer(long idLong) {
-        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer((int) idLong, false, null);
+        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer(idLong, false, null);
         if (!playerStats.isEmpty()) {
             HashMap<LocalDateTime, BigDecimal> results = new HashMap<>();
 
@@ -578,7 +578,7 @@ public class ChartServiceImpl implements IChartService {
     }
 
     private HashMap<Integer, BigDecimal> getSRStatsForPlayerForComparison(long idLong) {
-        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer((int) idLong, false, null);
+        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer(idLong, false, null);
         if (!playerStats.isEmpty()) {
             HashMap<Integer, BigDecimal> results = new HashMap<>();
 
@@ -627,7 +627,7 @@ public class ChartServiceImpl implements IChartService {
 
     @Override
     public HashMap<LocalDateTime, BigDecimal> getMedStatsForPlayer(long idLong) {
-        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer((int) idLong, false, null);
+        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer(idLong, false, null);
         if (!playerStats.isEmpty()) {
             HashMap<LocalDateTime, BigDecimal> results = new HashMap<>();
 
@@ -685,7 +685,7 @@ public class ChartServiceImpl implements IChartService {
 
     @Override
     public HashMap<LocalDateTime, Integer> getKLStatsForPlayer(long idLong) {
-        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer((int) idLong, false, null);
+        List<PlayerStatDTO> playerStats = globalDao.getStatsForPlayer(idLong, false, null);
 
         if (!playerStats.isEmpty()) {
             HashMap<LocalDateTime, Integer> results = new HashMap<>();
